% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{aggMatrix}
\alias{aggMatrix}
\title{Aggregate Values Accross Rows of a Matrix}
\usage{
aggMatrix(x, index, fun = sum, margin = 1, ...)
}
\arguments{
\item{x}{a matrix containing numeric, complex, integer or logical values,
or a  data frame.}

\item{index}{list of one or more factors, each of same length as
\code{dim(x)[1]} (as in \code{tapply}). The elements are coerced
to factors by as.factor.}

\item{fun}{the function to be applied. In the case of functions
like \code{+}, \code{\%*\%}, etc., the function name must be backquoted
or quoted.}

\item{margin}{over which margin should the operation take place
(i.e. which margin matches \code{index}).}

\item{\dots}{other arguments to be passed to \code{fun}.}
}
\value{
a matrix with dimensions \code{dim(x)[1], length(index)}
}
\description{
Apply a function columnwise to groups of rows in a matrix,
 according to a grouping index. Essentially applies the same
 \code{\link{tapply}} to multiple colums of a matrix.
}
\examples{

x <- expand(rnorm(5), 3)
aggMatrix(x, c(1, 1, 1, 2, 2))

}
\seealso{
Other utility: \code{\link{badRows}},
  \code{\link{cellIdx}}, \code{\link{expand}},
  \code{\link{merge.formula}}, \code{\link{rowProds}}
}

