% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{badRows}
\alias{badRows}
\title{Find rows with NA values in a dataframe or matrix}
\usage{
badRows(x)
}
\arguments{
\item{x}{a dataframe or matrix}
}
\value{
a logical vector
}
\description{
For each row, if any element is NA TRUE is returned.
}
\examples{
# fake data matrix with a missing row
x <- matrix(rnorm(3 * 4), nrow = 3)
x[2, 3] <- NA
badRows(x)
}
\seealso{
Other utility: \code{\link{+.formula}},
  \code{\link{merge.formula}}; \code{\link{aggMatrix}};
  \code{\link{cellIdx}}; \code{\link{expand}};
  \code{\link{rowProds}}
}

