% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{bufferMask}
\alias{bufferMask}
\title{Create a mask by distance from points}
\usage{
bufferMask(mask, points, buffer = 0)
}
\arguments{
\item{mask}{a \code{Raster*} object, used as the starting mask. All cells which
 are NA in \code{mask} will be NA in the resulting \code{RasterLayer}.}

\item{points}{a set of points (given as a two-column matrix of coordinates)
 around which top calculate the buffer.}

\item{buffer}{the distance from points beyond which to set cells to NA.
 If the coordinate system of \code{mask} is latitude/longitude, this will be
 in metres, otherwise in the units of the \code{mask}'s coordinate system.}
}
\value{
A \code{RasterLayer} with the same extent, resolution and coordinate system
as \code{mask} and with non-NA cells being 0.
}
\description{
Create a mask \code{RasterLayer} with NA values in all cells more
 than a specified distance from a set of points.
}
\seealso{
Other GIS: \code{\link{insertRaster}}
}

