% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utility.R
\name{cellIdx}
\alias{cellIdx}
\title{Find non-NA Raster or RasterBrick Cell numbers}
\usage{
cellIdx(x)
}
\arguments{
\item{x}{a Raster or RasterBrick object}
}
\value{
a numeric vector foi cell numbers
}
\description{
Get the cell numbers matching all non-NA cells.
}
\seealso{
Other utility: \code{\link{badRows}}
}

