% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demography_functions.R
\name{Date2cmc}
\alias{Date2cmc}
\alias{cmc2Date}
\alias{cmc2year}
\title{convert between Date and cmc formats}
\usage{
Date2cmc(Date)

cmc2Date(cmc, day = 1)

cmc2year(cmc)
}
\arguments{
\item{Date}{an object of class \code{Date} to convert to century-month-code
format.}

\item{cmc}{an object of class \code{numeric} representing a month in
century-month-code to convert to a \code{Date}}

\item{day}{the day of the month to assign when creating the date}
}
\value{
an object of class \code{numeric} (for \code{Date2cmc}) of of class
 \code{Date} (for \code{cmc2Date}).
}
\description{
Utility functions to convert between R's \code{Date} class and
  century-month-code (cmc) format, which measures time as month integers
}

