% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{expand}
\alias{expand}
\title{Expand a Vector into a Multi-column Matrix}
\usage{
expand(x, n)
}
\arguments{
\item{x}{a vector (data to repeat into \code{n} columns)}

\item{n}{an integer (number of times to repeat \code{x})}
}
\value{
a matrix with dimensions \code{length(x), n}, with \code{x} repeated
 in each column
}
\description{
Create a matrix by combining multiple copies of a vector as
 columns
}
\examples{

x <- rnorm(5)
expand(x, 3)

}
\seealso{
Other utility: \code{\link{aggMatrix}},
  \code{\link{badRows}}, \code{\link{cellIdx}},
  \code{\link{merge.formula}}, \code{\link{rowProds}}
}

