% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{getArea}
\alias{getArea}
\title{Return the Area of an sp Object}
\usage{
getArea(sp)
}
\arguments{
\item{sp}{an object of class \code{sp} represnting one or
more polygons}
}
\value{
a scalar numeric giving the total area
}
\description{
Determine the area of an sp polygon object,
 summing across sub-polygons if necesssary
}
\seealso{
Other GIS: \code{\link{bufferMask}}, \code{\link{condSim}},
  \code{\link{getPoints}}, \code{\link{insertRaster}},
  \code{\link{ll2cart}}, \code{\link{safeMask}}
}

