% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_functions.R
\name{getINLAParameters}
\alias{getINLAParameters}
\title{Extracting parameters from INLA MBG models}
\usage{
getINLAParameters(inla, method = c("sample", "MAP"), n = 1)
}
\arguments{
\item{inla}{a fitted \code{\link{inla}} object with fixed effects terms
and a spatial or spatio-temporal random effect model. If
\code{method = "sample"} then \code{inla} must have been fitted with the
argument \code{control.compute = list(config=TRUE)}.}

\item{method}{whether to draw \code{n} samples from the predict posterior
(if \code{method = "sample"}, the default) or to make predictions at the
maximum \emph{a posteriori} estimates of the model parameters (if
\code{method = "MAP"}).  If \code{method = "sample"} then \code{inla} must
have been fitted with the argument
\code{control.compute = list(config=TRUE)}.}

\item{n}{if \code{method = "sample"}, the number of samples to draw from the
predictive posterior.}
}
\value{
An object of class \code{params}, being a list with three elements:
 \itemize{
   \item params parameters
   \item names parameter names
   \item indices
     \itemize{
       \item fixed_idx index to fixed effect parameters in params and names
       \item spatial_idx index to spatial random effect parameters in params and names
       \item hyper_idx index to hyperparameters in params and names
     }
 }
}
\description{
Given a fitted \code{\link{inla}} object for a spatial or
 spatio-temporal geostatistical model, extract model parameters
 either from the maximum \emph{a posterior} parameter set, or as samples
 from the predictive posterior.
}

