% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{insertRaster}
\alias{insertRaster}
\title{Insert new values into a Raster* template}
\usage{
insertRaster(raster, new_vals, idx = NULL)
}
\arguments{
\item{raster}{a Raster* object with non-NA cells in which to insert new_vals}

\item{new_vals}{a dataframe of new values to put in the non-NA cells of
\code{raster}, with values from each row in a separate layer}

\item{idx}{an optional index giving the cell numbers corresponding to the
rows of \code{new_vals}. If not provided, it will be calculated from
\code{raster} using \code{\link{cellIdx}}.}
}
\value{
a RasterBrick object with the same extent and resolution as
 \code{raster} and as many layers as the number of columns in
 \code{new_vals}.
}
\description{
Given a template raster, new set of values for non-missing
 cells and optional index for non-missing cells, create a Raster* object
 containing the new values. The raster object will be replicate to have
 as many layers as the new dataframe has columns.
}
\seealso{
Other GIS: \code{\link{bufferMask}}, \code{\link{condSim}},
  \code{\link{getArea}}, \code{\link{getPoints}},
  \code{\link{ll2cart}}, \code{\link{makeVoronoiPolygons}},
  \code{\link{safeMask}}, \code{\link{sortPolyData}}
}

