% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{ll2cart}
\alias{ll2cart}
\alias{cart2ll}
\title{Convert from Latitude/Longitude to Cartesian Coordinates}
\usage{
ll2cart(longlat, radius = 6371)

cart2ll(xyz, radius = 6371)
}
\arguments{
\item{longlat}{a matrix or dataframe giving the longitudes (first column)
and latitudes (second column) of a set of locations}

\item{radius}{radius of the sphere in on which the Cartesian coodinates are
defined. By default the approximate radius of the earth in kilometres.}

\item{xyz}{a matrix or dataframe giving the cartesian coordinates
(in order x, y, z) of a set of locations}
}
\value{
ll2cart: a three-column dataframe giving the x, y and z positions
 for each set of coordinates

cart2ll: a two-column dataframe giving the longitudes and latitudes
 for each set of coordinates
}
\description{
Convert coordinates given as latitude and longitude to
 three-dimensional coordinates on a sphere. This can be used to account
 for great-circle distances when fitting spatial models.
}
\seealso{
Other GIS: \code{\link{bufferMask}}, \code{\link{condSim}},
  \code{\link{getArea}}, \code{\link{getPoints}},
  \code{\link{insertRaster}},
  \code{\link{makeVoronoiPolygons}},
  \code{\link{safeMask}}, \code{\link{sortPolyData}}

Other GIS: \code{\link{bufferMask}}, \code{\link{condSim}},
  \code{\link{getArea}}, \code{\link{getPoints}},
  \code{\link{insertRaster}},
  \code{\link{makeVoronoiPolygons}},
  \code{\link{safeMask}}, \code{\link{sortPolyData}}
}
