% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{makeVoronoiPolygons}
\alias{makeVoronoiPolygons}
\title{Make Voronoi polygons from an INLA mesh and boundary}
\usage{
makeVoronoiPolygons(mesh, boundary)
}
\arguments{
\item{mesh}{an \code{inla.mesh} object.}

\item{boundary}{an \code{inla.mesh.segment} object used to specify the Voronoi polygons boundary.}
}
\value{
An object of \code{SpatialPolygons} class.
}
\description{
Create Voronoi polygons based on an \code{inla.mesh} object
and an \code{inla.mesh.segment} object specifying the boundary (edge) of the Voronoi polygons.
}
\examples{
# load packages
library(sp)
library(INLA)

# make a SpatialPolygons object
poly <- SpatialPolygons(list(
Polygons(list(Polygon(matrix(c(76, 35, 90, 34, 60, 20, 50, 31), ncol=2, byrow=TRUE))), ID=1)))

# make mesh
mesh <- inla.mesh.2d(
  boundary = inla.sp2segment(poly),
  max.edge = c(3,5))

# make voronoi polygons
v <- makeVoronoiPolygons(mesh, poly)
}
\seealso{
Other GIS: \code{\link{bufferMask}}; \code{\link{cart2ll}},
  \code{\link{ll2cart}}; \code{\link{condSim}};
  \code{\link{getArea}}; \code{\link{getPoints}};
  \code{\link{insertRaster}}; \code{\link{safeMask}}
}

