% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{merge.formula}
\alias{+.formula}
\alias{merge.formula}
\title{Merge Two Formulae}
\usage{
\method{merge}{formula}(x, y, ...)

\method{+}{formula}(x, y)
}
\arguments{
\item{x}{a \code{formula} object with both response and predictor terms}

\item{y}{a \code{formula} object with predictor terms (response ignored)}

\item{\dots}{further arguments passed to or from other methods.}
}
\value{
a \code{formula} with the response form \code{x} and the predictors
 from both \code{x} and \code{y}
}
\description{
Merge two formulae to combine predictors (right hand side).
}
\examples{
f <- y ~ x
g <- ~ a * b + c

fg <- merge(f, g)
fg



fg <- f + g
fg

}
\seealso{
Other utility: \code{\link{aggMatrix}},
  \code{\link{badRows}}, \code{\link{cellIdx}},
  \code{\link{expand}}, \code{\link{rowProds}}
}

