% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_functions.R
\name{pcaTrans}
\alias{pcaTrans}
\title{PCA transform a \code{Raster*} object}
\usage{
pcaTrans(coords, covs)
}
\arguments{
\item{coords}{a two-column matrix or dataframe giving the location about
which to carry out the PCA analysis}

\item{covs}{a \code{Raster*} object containing covariates to rotate}
}
\value{
a \code{Raster*} object with the same extent, resolutiona and number
 of layers as \code{covs} but with each layer giving the location on a
 different principal component axis.
}
\description{
Given a matrix of point coordinates and a Raster* object of
 covariates, return a \code{Raster*} object of the same size with the layers
 giving principal components from a PCA rotation based on the data values
 at the coordinates.
}
