% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demography_functions.R
\name{periodTabulate}
\alias{periodTabulate}
\title{Tabulate Exposures/Deaths from CBH records}
\usage{
periodTabulate(age_death, birth_int, cluster_id, windows_lower = c(0, 1, 3, 6,
  12, 24, 36, 48), windows_upper = c(0, 2, 5, 11, 23, 35, 47, 59),
  period = 60, nperiod = 1, delay = max(windows_upper - windows_lower))
}
\arguments{
\item{age_death}{numeric vector giving the age of death in months of each
child. Note that if the child is still alive at the time of interview,
corresponding eleents of \code{age_death} should be given a number greater
than \code{max(windows_upper)}}

\item{birth_int}{numeric vector giving the time in month between the
child's birth and the interview}

\item{cluster_id}{vector giving the cluster identifier (can be numeric,
cahracter or factor)}

\item{windows_lower,windows_upper}{numeric vectors giving the
non-overlapping lower and upper ages in months of each survival window}

\item{period}{the length of time in months for which mortality rates
should be estimated, either vector or scalar}

\item{nperiod}{the number of consecutive periods to calculate for. I.e.
if \code{period = 12} and \code{nperiod = 3}, numbers will be returned for
periods 0-12, 13-24 and 25-36 months prior to the interview (plus delay).}

\item{delay}{the length of time in months prior to the interview date
 to end the period, either vector or scalar.
 I.e. the period runs from \code{period + delay}
 months to \code{delay} months before the interview month.}
}
\value{
a dataframe with number of rows equal to all combinations of
 clusters, age bins and periods; i.e.
 \code{length(unique(cluster_id)) * length(windows_lower) * nperiod},
 and four columns:
 \itemize{
   \item \code{cluster_id} identifier for the cluster
   \item \code{exposed} number exposed in a given period and age bin
   \item \code{died} number exposed and died in a given period and age bin
   \item \code{period} the corresponding period number
   \item \code{age_bin} the corresponding age bin number
 }
}
\description{
Count the number exposed and number dying
 in a given set of age ranges and time periods
}

