% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_functions.R
\name{predictAll}
\alias{predictAll}
\title{Predict the Linear Predictor of an INLA MBG Model}
\usage{
predictAll(draw, params, mesh, coords, data, fixed_subset = NULL,
  fixed = TRUE, spatial = TRUE)
}
\arguments{
\item{draw}{which parameter set to use}

\item{params}{an object of class \code{params} produced by
\code{\link{getINLAParameters}}.}

\item{mesh}{the \code{\link{inla.mesh}} object used to construct the spatial
random effect.}

\item{coords}{a two-column dataframe with the coordinates to evaluate the
spatial random effect at.}

\item{data}{a dataframe with all the columns referred to in \code{params}.}

\item{fixed_subset}{an optional character vector giving a subset of covariates
to include in the fixed effects prediction.}

\item{fixed}{whether to include fixed effects terms in the predictor}

\item{spatial}{whether to include spatial terms in the predictor}
}
\value{
a vector of the same length as the number of rows in \code{coords}
 and \code{data}, giving the corresponding values of the spatial linear
 predictor (both from the fixed and reandom effects)
}
\description{
For a given set of parameters and new dataset, get the
 predicted values of the linear predictor form the fixed and spatial
 random effects.
}

