% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inla_functions.R
\name{predictFixed}
\alias{predictFixed}
\title{Predict the Fixed Effect Parts of an INLA MBG Model}
\usage{
predictFixed(params, data)
}
\arguments{
\item{params}{an object of class \code{params} produced by
\code{\link{getINLAParameters}}}

\item{data}{a dataframe with all the columns referred to in \code{params}}
}
\value{
a vector of the same length as the number of rows in \code{data},
 giving the values of the linear predictor corresponding to the records in
 \code{data}
}
\description{
For a given set of paramters and new dataset, get the
 component of the linear predictor corresponding to the fixed effects.
}

