% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inla_functions.R
\name{predictINLA}
\alias{predictINLA}
\title{Prediction from INLA MBG models}
\usage{
predictINLA(inla, data, mesh, coords = c("lat", "long"), type = c("link",
  "response"), method = c("sample", "MAP"), n = 1, ncpu = 1)
}
\arguments{
\item{inla}{a fitted \code{\link{inla}} object with fixed effects terms
and a spatial or spatio-temporal random effect model. If
\code{method = "sample"} then \code{inla} must have been fitted with the
argument \code{control.compute = list(config=TRUE)}.}

\item{data}{a dataframe giving all of the covariates for the fixed effects
part of the model and the temporal term is applicable.}

\item{mesh}{the \code{\link{inla.mesh}} object used to construct the spatial
random effect.}

\item{coords}{a character vector of length two giving the names of the
columns in \code{data} which contain the coordinates of locations at which
to make predictions.}

\item{type}{the type of prediction required. The default is on the scale of
the linear predictors; the alternative \code{"response"} is on the scale of
the response variable. Thus for a default binomial model the default
predictions are of log-odds (probabilities on logit scale) and
\code{type = "response"} gives the predicted probabilities.}

\item{method}{whether to draw \code{n} samples from the predict posterior
(if \code{method = "sample"}, the default) or to make predictions at the
maximum \emph{a posteriori} estimates of the model parameters (if
\code{method = "MAP"}).  If \code{method = "sample"} then \code{inla} must
have been fitted with the argument
\code{control.compute = list(config=TRUE)}.}

\item{n}{if \code{method = "sample"}, the number of samples to draw from the
predictive posterior.}

\item{ncpu}{the number of cores to use to make predictions. If
 \code{ncpu = 1} then predicitons will be made sequentially, otherwise for
 \code{n > 1} a snowfall cluster will be initiated and predictions run in
 parallel.}
}
\value{
a matrix with the same number of rows as \code{data} giving the
 predicted values at these locations
}
\description{
Given a fitted \code{\link{inla}} object for a spatial or
 spatio-temporal geostatistical model, make predictions to a new dataset
 either from the maximum \emph{a posterior} parameter set, or as samples
 from the predictive posterior.
}

