% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{rowProds}
\alias{rowProds}
\title{Form Row Products}
\usage{
rowProds(x, na.rm = FALSE, dims = 1)
}
\arguments{
\item{x}{an array of two or more dimensions, containing numeric, complex,
integer or logical values, or a numeric data frame.}

\item{na.rm}{logical. Should missing values (including NaN) be omitted from
the calculations?}

\item{dims}{integer: Which dimensions are regarded as rows to
sum over. The product is over dimensions dims+1.}
}
\value{
A numeric or complex array of suitable size, or a vector if the
 result is one-dimensional.
}
\description{
Form row products for numeric arrays
}
\examples{

x <- expand(rnorm(5), 3)
rowProds(x)

}
\seealso{
Other utility: \code{\link{aggMatrix}},
  \code{\link{badRows}}, \code{\link{cellIdx}},
  \code{\link{expand}}, \code{\link{merge.formula}}
}
