% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{safeMask}
\alias{safeMask}
\title{safeMask}
\usage{
safeMask(raster, sp)
}
\arguments{
\item{raster}{a \code{RasterLayer} object to mask.}

\item{sp}{an \code{sp} object by which to mask \code{raster}.}
}
\value{
a \code{RasterLayer} object, the same as \code{raster} but with
 any cell falling under \code{sp} set to \code{NA}
}
\description{
Mask a \code{RasterLayer} object using a
 \code{sp} object representing one or more polygons in a
 such that cells falling under small polygons are masked out.
 Note this function is slower than \code{raster::mask}
 and has different behaviour.
}
\seealso{
Other GIS: \code{\link{bufferMask}}, \code{\link{condSim}},
  \code{\link{getArea}}, \code{\link{getPoints}},
  \code{\link{insertRaster}}, \code{\link{ll2cart}},
  \code{\link{makeVoronoiPolygons}},
  \code{\link{sortPolyData}}
}

