% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{sortPolyData}
\alias{sortPolyData}
\title{Organise polygon count data for modelling.}
\usage{
sortPolyData(polygons, data, covariates, fun = rep("mean",
  nlayers(covariates)))
}
\arguments{
\item{polygons}{an object of \code{SpatialPolygons} class.}

\item{data}{a \code{data.frame} containing at least columns Longitude and Latitude}

\item{covariates}{a \code{RasterStack}}

\item{fun}{vector of length equalling the number of covariates specifying the function to be
applied across pixel values within each polygon for each covariate. The default is 'mean'. 
Functions must be entered as strings and 'modal' should be specified for discrete variables.}
}
\value{
A \code{data.frame} of, for each polygon, centroid coordinates, 
counts of occurrence points and a summary of covariate values.
}
\description{
Takes polygons, an occurrence dataset and covariates. Returns a \code{data.frame} containing
 centroid coordinates, count of occurrence points and a summary of covariate values for each polygon. 
 Includes an option to specify the function to apply across pixel values for each covariate. 
 The default is mean.
}
\examples{
# load packages
library(sp)
library(raster)
library(seegSDM)

# make voronoi polygons
example(makeVoronoiPolygons)

# make occcurrence data
occ <- matrix(c(-5, 0,
                 -1, 6,
                 -8, 8,
                 -8, 7,
                 -5, 10),
             ncol=2,
             byrow=TRUE)

# convert to dataframe
occ <- as.data.frame(occ)

# add Longitude and Latitude columns
colnames(occ)[colnames(occ)=='V1'] <- 'Longitude'
colnames(occ)[colnames(occ)=='V2'] <- 'Latitude'

# load covariates from seegSDM
data(covariates)

# make functions vector
fun = c('mean', 'mean', 'modal')

# sort data
dat <- sortPolyData(polygons = v, data = occ, covariates=covariates, fun=fun)
}
\seealso{
Other GIS: \code{\link{bufferMask}}, \code{\link{condSim}},
  \code{\link{getArea}}, \code{\link{getPoints}},
  \code{\link{insertRaster}}, \code{\link{ll2cart}},
  \code{\link{makeVoronoiPolygons}}, \code{\link{safeMask}}
}

