% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{cellIdx}
\alias{cellIdx}
\title{Find non-NA \code{Raster} or \code{RasterBrick} cell numbers}
\usage{
cellIdx(x)
}
\arguments{
\item{x}{a \code{Raster} or \code{RasterBrick} object}
}
\value{
a numeric vector of cell numbers
}
\description{
Get the cell numbers matching all non-NA cells.
}
\seealso{
Other utility: \code{\link{aggMatrix}},
  \code{\link{badRows}}, \code{\link{expand}},
  \code{\link{merge.formula}}, \code{\link{rowProds}}
}

