% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{condSim}
\alias{condSim}
\title{condSim}
\usage{
condSim(vals, weights = NULL, group = NULL, fun = sum)
}
\arguments{
\item{vals}{a matrix of samples of pixel-level vales where each row
corresponds to a different pixels and each column to a different posterior
sample}

\item{weights}{an optional vector of weights in each cell, corresponding to
the rows of \code{vals}.}

\item{group}{an optional vector (of the same size as \code{weights})
identifying the group (e.g. an admin unit) to which each pixel belongs. If
specified, samples of case counts are calculated for each unique value of
\code{group}. If \code{group = NULL}, samples are returned as the total
over all pixels.}

\item{fun}{to apply to summarise the (weighted) elements of \code{vals},
within each group, for each sample.}
}
\value{
If \code{group = NULL}, a vector of size equal to the number of
 columns in \code{vals}, each element giving a different simulated summary
 across all pixels covered by \code{vals}.
 If \code{group} is specified, a matrix of simulated summaries with each
  row corresponding to a different unique value in \code{group} (e.g. an
   administrative unit) and each column corresponding to a different draw.
}
\description{
Carry out conditional simulation on a matrix of pixel-level
  value simulations to calculate overall or regional estimates of metrics,
  such total case numbers, prevalences or inequality metrics
}
\examples{
# make some fake prevalence map data
n_pixels <- 100
n_draws <- 10
prevalence <- matrix(runif(n_pixels * n_draws),
                     ncol = n_draws)
population <- rpois(n_pixels, 100)

# run overall simulation
draws <- condSim(prevalence, population)

# simulate by (made up) country to get the expected number of infections
country <- sample(letters[1:5], n_pixels, replace = TRUE)
draws <- condSim(prevalence, population, country)

}
\seealso{
Other GIS: \code{\link{bufferMask}}, \code{\link{getArea}},
  \code{\link{getPoints}}, \code{\link{insertRaster}},
  \code{\link{ll2cart}}, \code{\link{safeMask}}
}

