% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gis_functions.R
\name{getPoints}
\alias{getPoints}
\title{getPoints}
\usage{
getPoints(shape, raster, n = 10, prob = FALSE)
}
\arguments{
\item{shape}{an \code{sp} object containing one or more polygons
representing the region for which integration points are required}

\item{raster}{a \code{RasterLayer} object, optionally containing values
determining weights for integration}

\item{n}{the number of integration points required}

\item{prob}{whether to weight the integration points by the values of
 \code{raster}}
}
\value{
a three-column matrix giving the coordinates and corresponding
 weights for the spatial integration points over \code{sp}.
}
\description{
Use random sampling and K-means clustering to generate
 a set of coordinates and corresponding weights that can be used
 to carry out spatial integration over an area represented by an
 \code{sp} object. The weights can be purely spatial, or can
 determined by the values of a raster representing the process of
 interest - for example population density. The raster should be
 such that random points can be sampled from it according to cell
 values using \code{seegSDM::bgSample}.
}
\seealso{
Other GIS: \code{\link{bufferMask}}; \code{\link{cart2ll}},
  \code{\link{ll2cart}}; \code{\link{getArea}};
  \code{\link{insertRaster}}; \code{\link{safeMask}}
}

