% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/demography_functions.R
\name{periodMortality}
\alias{periodMortality}
\title{Synthetic Cohort Life Table Child Mortality Estimation}
\usage{
periodMortality(age_death, birth_int, cluster_id, windows_lower = c(0, 1, 3,
  6, 12, 24, 36, 48), windows_upper = c(0, 2, 5, 11, 23, 35, 47, 59),
  ages_lower = c(0, 0), ages_upper = c(12, 60), period = 60,
  delay = max(windows_upper - windows_lower), glm = FALSE)
}
\arguments{
\item{age_death}{numeric vector giving the age of death in months of each
child. Note that if the child is still alive at the time of interview,
corresponding eleents of \code{age_death} should be given a number greater
than \code{max(windows_upper)}}

\item{birth_int}{numeric vector giving the time in month between the
child's birth and the interview}

\item{cluster_id}{vector giving the cluster identifier (can be numeric,
cahracter or factor)}

\item{windows_lower,windows_upper}{numeric vectors giving the
non-overlapping lower and upper ages in months of each survival window}

\item{ages_lower,ages_upper}{numeric vectors giving the
non-overlapping lower and upper ages in months of the age windows for
which to estimate mortality rates}

\item{period}{the length of time in months for which mortality rates
should be estimated, either vector or scalar}

\item{delay}{the length of time in months prior to the interview date
to end the period, either vector or scalar.
I.e. the period runs from \code{period + delay}
months to \code{delay} months before the interview month.}

\item{glm}{whether to infer the window-specific survival probabilities
 using a binomial random effects model across cluster, window and cohort.
 If \code{FALSE} probabilities are calculated as the raw ratio of the
 number that survived to the number exposed and may therefore contain
 zeros and indeterminate values.}
}
\value{
a list with the same number of elements as \code{ages}, each
 element being a dataframe containing the estimated mortality rates
 and effective sample size (weighted number exposed in each suvival
window).
}
\description{
Compute child mortality estimates using the DHS three-cohort
 life table approach.
}

