% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_functions.R
\name{predictINLA}
\alias{predictINLA}
\alias{predictRasterINLA}
\title{Prediction from INLA MBG models}
\usage{
predictINLA(inla, data, mesh, coords = c("Longitude", "Latitude"),
  type = c("link", "response"), method = c("sample", "MAP"), n = 1,
  fixed = TRUE, spatial = TRUE, fixed_subset = NULL, ncpu = 1)

predictRasterINLA(inla, raster, mesh, constants = list(), ...)
}
\arguments{
\item{inla}{a fitted \code{\link{inla}} object with fixed effects terms
and a spatial or spatio-temporal random effect model. If
\code{method = "sample"} then \code{inla} must have been fitted with the
argument \code{control.compute = list(config=TRUE)}.}

\item{data}{a dataframe giving all of the covariates for the fixed effects
part of the model and the temporal term is applicable.}

\item{mesh}{the \code{\link{inla.mesh}} object used to construct the spatial
random effect.}

\item{coords}{a character vector of length two giving the names of the
columns in \code{data} which contain the coordinates of locations at which
to make predictions.}

\item{type}{the type of prediction required. The default is on the scale of
the linear predictors; the alternative \code{"response"} is on the scale of
the response variable. Thus for a default binomial model the default
predictions are of log-odds (probabilities on logit scale) and
\code{type = "response"} gives the predicted probabilities.}

\item{method}{whether to draw \code{n} samples from the predict posterior
(if \code{method = "sample"}, the default) or to make predictions at the
maximum \emph{a posteriori} estimates of the model parameters (if
\code{method = "MAP"}).  If \code{method = "sample"} then \code{inla} must
have been fitted with the argument
\code{control.compute = list(config=TRUE)}.}

\item{n}{if \code{method = "sample"}, the number of samples to draw from the
predictive posterior.}

\item{fixed}{whether to include fixed effects terms in the predictor}

\item{spatial}{whether to include spatial terms in the predictor}

\item{fixed_subset}{an optional character vector giving a subset of covariates
to include in the fixed effects prediction.}

\item{ncpu}{the number of cores to use to make predictions. If
\code{ncpu = 1} then predicitons will be made sequentially, otherwise for
\code{n > 1} a snowfall cluster will be initiated and predictions run in
parallel.}

\item{raster}{a \code{Raster*} object containing the fixed effects
covariates to use for prediction.}

\item{constants}{an optional named list giving constant values for named
fixed effects in \code{inla}. E.g. for an intercept term names \code{int},
do \code{constants = list(int = 1)}.}

\item{\dots}{arguments to be passed to \code{predictINLA}.}
}
\value{
predictINLA: a matrix with the same number of rows as \code{data} giving the
 predicted values at these locations

predictRasterINLA: a \code{RasterBrick} or \code{RasterStack}
 (if \code{method = 'sample'}) or \code{RasterLayer} (if
 \code{method = 'MAP'}) giving pixel-level predictions from \code{inla}.
}
\description{
Given a fitted \code{\link{inla}} object for a spatial
 geostatistical model, make predictions to a new dataset
 or \code{RasterBrick} either from the maximum \emph{a posterior} parameter
 set, or as samples from the predictive posterior.

 Note that this function is only designed to work with a specific type of
 geostatistical model, fitted in a specific way and is likely to produce
 unexpected (i.e. wrong) results when used with other types of model.

 Examples of things that aren't allowed are:
   \itemize{
     \item prediction using factor (discrete) variables.
      Though it should be possible to manually expand the factors into the
      correctly named dummy variables (see \code{inla$names.fixed} for the
      expected names) and predict to these. This includes the use of iid
      random effects, unfortunately.
     \item prediction using models constructed with anything other than an
      spde2 spatial model.
     \item spatial models fitted using 3D Cartesian coordinates (to be
      added soon).
     \item space-time models (to be added soon).
   }
}

