% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/inla_functions.R
\name{predictSpatial}
\alias{predictSpatial}
\title{Predict the Spatial Random Effect Part of an INLA MBG Model}
\usage{
predictSpatial(params, coords, mesh, draw = 1, time = NULL, n_time = NULL)
}
\arguments{
\item{params}{an object of class \code{params} produced by
\code{\link{getINLAParameters}}.}

\item{coords}{a two-column dataframe with the coordinates to evaluate the
spatial random effect at.}

\item{mesh}{the \code{\link{inla.mesh}} object used to construct the spatial
random effect.}

\item{draw}{which parameter set to use}

\item{time}{if a space-time model, the index to the discrete time variable
for which spatial predictions are to be made in this round.}

\item{n_time}{if a space-time model, the length of the discrete time
 variable used in the model.}
}
\value{
a vector of the same length as the number of rows in \code{coords},
 giving the corresponding values of the spatial linear predictor
}
\description{
For a given set of paramters and new set of coordinates,
 get the component of the linear predictor corresponding to the spatial
 random effect, by linear interpolation.
}

